reorganize 30_days folder
