+++ 30_days/d10_binary_numbers/binary_numbers.cbp
