+++ 30_days/d10_binary_numbers/consecutive_one.c
