+++ 30_days/d8_dic_and_maps/out.txt
